# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "МедМаршрут API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/route/start": {
      "post": {
        "summary": "Start Route",
        "operationId": "start_route_api_v1_route_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reminders": {
      "get": {
        "summary": "Reminders",
        "operationId": "reminders_api_v1_reminders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReminderDto"
                  },
                  "type": "array",
                  "title": "Response Reminders Api V1 Reminders Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Reminder",
        "operationId": "add_reminder_api_v1_reminders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reminders/{reminder_id}/toggle": {
      "post": {
        "summary": "Toggle Reminder",
        "operationId": "toggle_reminder_api_v1_reminders__reminder_id__toggle_post",
        "parameters": [
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Reminder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meds": {
      "get": {
        "summary": "Meds",
        "operationId": "meds_api_v1_meds_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedsOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Med",
        "operationId": "add_med_api_v1_meds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedAddIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meds/{course_id}/intake-toggle": {
      "post": {
        "summary": "Toggle Intake",
        "operationId": "toggle_intake_api_v1_meds__course_id__intake_toggle_post",
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MedSlotDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CustomStepIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "deadline_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Time"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CustomStepIn",
        "description": "Свой шаг при создании маршрута из конструктора."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MedAddIn": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "times": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Times"
          },
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          }
        },
        "type": "object",
        "required": [
          "name",
          "times"
        ],
        "title": "MedAddIn"
      },
      "MedCourseDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "times": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Times"
          },
          "until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Until"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "times",
          "enabled"
        ],
        "title": "MedCourseDto"
      },
      "MedSlotDto": {
        "properties": {
          "course_id": {
            "type": "integer",
            "title": "Course Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "at_time": {
            "type": "string",
            "title": "At Time"
          },
          "taken": {
            "type": "boolean",
            "title": "Taken"
          },
          "taken_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taken At"
          }
        },
        "type": "object",
        "required": [
          "course_id",
          "name",
          "at_time",
          "taken"
        ],
        "title": "MedSlotDto"
      },
      "MedsOut": {
        "properties": {
          "courses": {
            "items": {
              "$ref": "#/components/schemas/MedCourseDto"
            },
            "type": "array",
            "title": "Courses"
          },
          "today": {
            "items": {
              "$ref": "#/components/schemas/MedSlotDto"
            },
            "type": "array",
            "title": "Today"
          }
        },
        "type": "object",
        "required": [
          "courses",
          "today"
        ],
        "title": "MedsOut"
      },
...
      "ReminderDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "place": {
            "type": "string",
            "title": "Place"
          },
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "done_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Done At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "place",
          "at",
          "enabled"
        ],
        "title": "ReminderDto"
      },
...
      "RouteDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "situation_key": {
            "type": "string",
            "title": "Situation Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "subtitle": {
            "type": "string",
            "title": "Subtitle"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "total_steps": {
            "type": "integer",
            "title": "Total Steps"
          },
          "done_steps": {
            "type": "integer",
            "title": "Done Steps"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepDto"
            },
            "type": "array",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "id",
          "situation_key",
          "title",
          "subtitle",
          "active",
          "total_steps",
          "done_steps",
          "steps"
        ],
        "title": "RouteDto"
      },
      "RouteStartIn": {
        "properties": {
          "situation_key": {
            "type": "string",
            "title": "Situation Key"
          },
          "exclude_positions": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Exclude Positions"
          },
          "custom_steps": {
            "items": {
              "$ref": "#/components/schemas/CustomStepIn"
            },
            "type": "array",
            "title": "Custom Steps"
          }
        },
        "type": "object",
        "required": [
          "situation_key"
        ],
        "title": "RouteStartIn"
      },
...
      "StepDto": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "place": {
            "type": "string",
            "title": "Place"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "deadline_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "has_checklist": {
            "type": "boolean",
            "title": "Has Checklist"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "template"
          }
        },
        "type": "object",
        "required": [
          "id",
          "position",
          "title",
          "description",
          "place",
          "status",
          "has_checklist"
        ],
        "title": "StepDto"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}